rename: 패키지 명 sixPieces->pieces
