+++ src/main/java/softeer2nd/chess/Board.java
@@ -1,6 +1,6 @@
 package softeer2nd.chess;
 
-import softeer2nd.chess.sixPieces.Pawn;
+import softeer2nd.chess.pieces.Pawn;
 
 import java.util.ArrayList;
