+++ src/main/java/softeer2nd/chess/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.sixPieces;
+package softeer2nd.chess.pieces;
 
 public class Pawn {
